Updated adder tree and control
Adder tree should be much better now. As of now it takes a single data stream in , we need to figure out if it'll take 9 inputs or one that has all 9. Control is updated a bit, we need to define the states and how the convolver should act in each state explicitly